requirements.txt
psycopg2-binary
typer
python-dotenv